--- examples/LandmarkDetection/DQN/CardiacLocalBaseline
@@ -1,31 +1,0 @@
-14AB01963,42.028003999942946,10.63014581273465
-14AB02212,-6.386617026583323,36.61966684720111
-14AB02563,51.90170470853913,19.209372712298546
-14AC02307,32.77237063065526,49.68903299521938
-14AD01860,17.872615271639038,34.84250278036869
-14AE02363,76.48368462882516,13.152946437965905
-14AF03209,27.116027517814963,15.652475842498529
-14AG02058,55.40128139308628,10.862780491200215
-14AH01539,44.51180009962144,16.401219466856727
-14AH03190,47.10566520060208,33.18132004607411
-14AJ02284,-91.58607399935944,177.14118662806794
-14AL01844,5.428554193280874,55.036351623268054
-14AM01672,59.265447531215926,21.656407827707714
-14AM03052,39.15348965950637,36.742346141747674
-14AM03221,75.06635316238794,22.627416997969522
-14AP01771,13.176545131155265,47.75981574503821
-14AP02565,77.1921245590581,18.627936010197157
-14AR03162,104.28021406134822,1.4142135623730951
-14AS01700,-31.82094259871029,88.61715409558128
-14AS02334,77.63895166074333,2.8284271247461903
-14AS03134,82.7903099556195,27.313000567495326
-14AV01314,47.55851708044101,15.684387141358123
-14AW02659,68.61289549566999,9.219544457292887
-14BC01859,97.0366577817544,19.209372712298546
-14BD02466,61.22235199489571,28.0178514522438
-14BH01520,57.092328338389024,23.430749027719962
-14BJ02261,-101.61860251237142,176.21010186706096
-14BO02653,93.46262690369032,23.280893453645632
-14BP03151,34.09583278097686,13.892443989449804
-14BV02538,71.91996672816812,23.345235059857504
-14CB01779,40.336298282220426,46.69047011971501
